+++ _includes/live-demos/drive-standalone-start/example.js
@@ -1,0 +1,6 @@
+(function () {
+  tinydrive.start({
+    target: 'div#drive-standalone-start',
+    token_provider: '/your-server-endpoint/jwt-provider'
+  });
+})();